requirements.txt
pandas>=2.2.1
requests>=2.31.0
requests-cache>=1.2.0
python-dateutil>=1.16.0
tqdm>=4.66.2
beautifulsoup4>=4.12.3
openpyxl>=3.1.5
feature-engine>=1.8.2
catboost>=1.2.7
venn-abers>=1.4.6
joblib>=1.4.2
optuna>=4.1.0
pyarrow>=18.0.0
scikit-learn>=1.5.2
skfolio>=0.5.1
quantstats>=0.0.64
fullmonte>=0.0.4
ipython>=8.29.0
